Incorporated refactored changes for the Terms and Conditions Modal
